src/empty-unknown-url-db/requirements.txt
mongo
ramda
boto3